+++ src/main/java/io/paradaux/friendlybot/commands/utility/PingCommand.java
@@ -2,7 +2,7 @@
  * MIT License
  *
  * Copyright (c) 2021 Rían Errity
- * io.paradaux.friendlybot.commands.fun.PingCommand :  31/01/2021, 01:28
+ * io.paradaux.friendlybot.commands.utility.PingCommand :  31/01/2021, 01:32
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -23,7 +23,7 @@
  * SOFTWARE.
  */
 
-package io.paradaux.friendlybot.commands.fun;
+package io.paradaux.friendlybot.commands.utility;
 
 import com.jagrosh.jdautilities.command.CommandEvent;
 import io.paradaux.friendlybot.FriendlyBot;
